* changed pk + fk in db to int(10)

+++ config/db/tables.sql
@@ -3,10 +3,10 @@
 --
 
 CREATE TABLE `ext_sysmanager_extension` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` int(10) NOT NULL AUTO_INCREMENT,
   `date_create` int(10) unsigned NOT NULL default '0',
   `date_update` int(10) unsigned NOT NULL,
-  `ext` smallint(5) unsigned NOT NULL,
+  `ext` int(10) unsigned NOT NULL,
   `version` varchar(16) NOT NULL,
   PRIMARY KEY  (`id`)
 ) ENGINE=MyISAM  DEFAULT CHARSET=utf8;
